storage: use 64 bits for indexed physical offsets
A hard segment size limit was not enforced and some users created
segments larger than 4gb resulting in an overflow of the 32-bit physical
offset entry in the segment offset index.

The result of this was that readers would eventually seek backwards in a
segment to a non-aligned location and experience a checksum failure on
the header.

This PR switches to a 64-bit representation for the physical offset,
bumps the on-disk file format version for the index, and forces a
rebuild for older versions. This allows us to also remove code for
supporting the older formats.

+++ src/v/storage/segment_index.cc
@@ -25,7 +25,7 @@
 namespace storage {
 
 static inline segment_index::entry translate_index_entry(
-  const index_state& s, std::tuple<uint32_t, uint32_t, uint32_t> entry) {
+  const index_state& s, std::tuple<uint32_t, uint32_t, uint64_t> entry) {
     auto [relative_offset, relative_time, filepos] = entry;
     return segment_index::entry{
       .offset = model::offset(relative_offset + s.base_offset()),
